parallel execution refactoring
