--- examples/s3/basic/get_bucket_info.json
@@ -1,27 +1,0 @@
-{
-  "conversations": [
-    {
-      "host": "localhost:7480",
-      "auth": {
-        "access_key": "test",
-        "secret_key": "test"
-      },
-      "requests": [
-        {
-          "for": 1,
-          "auth": "aws_v2",
-          "method": "GET",
-          "uri": "admin/bucket",
-          "query_string": "format=json&bucket=foo",
-          "response": {
-            "out": {
-              "format": {
-                "body": "json"
-              }
-            }
-          }
-        }
-      ]
-    }
-  ]
-}
