--- examples/s3/basic/put_stat_bucket.json
@@ -1,41 +1,0 @@
-{
-  "conversations": [
-    {
-      "host": "localhost:7480",
-      "auth": {
-        "access_key": "test",
-        "secret_key": "test"
-      },
-      "requests": [
-        {
-          "for": 1,
-          "auth": "aws_v4",
-          "method": "PUT",
-          "uri": "foo",
-          "query_string": "format=json",
-          "response": {
-            "out": {
-              "format": {
-                "body": "json"
-              }
-            }
-          }
-        },
-        {
-          "for": 1,
-          "auth": "aws_v4",
-          "method": "HEAD",
-          "uri": "foo",
-          "query_string": "format=json",
-          "response": {
-            "out": {
-              "format": {
-                "body": "json"
-              }
-            }
-          }
-        }
-      ]
-    }
-  ]
-}
