--- examples/s3/basic/put_user.json
@@ -1,27 +1,0 @@
-{
-  "conversations": [
-    {
-      "host": "localhost:7480",
-      "auth": {
-        "access_key": "test",
-        "secret_key": "test"
-      },
-      "requests": [
-        {
-          "for": 1,
-          "auth": "aws_v2",
-          "method": "PUT",
-          "uri": "admin/user",
-          "query_string": "format=json&uid=jack&display-name=Jack",
-          "response": {
-            "out": {
-              "format": {
-                "body": "json"
-              }
-            }
-          }
-        }
-      ]
-    }
-  ]
-}
