--- examples/s3/scripted/test_include.js
@@ -1,13 +1,0 @@
-function DeclareVersion() {
-  return "v0.0.0";
-}
-
-let TLV = {
-  TRC: 0,
-  DBG: 1,
-  INF: 2,
-  WRN: 3,
-  ERR: 4,
-  CRI: 5,
-  OFF: 6,
-}
