--- examples/s3/scripted/test_scripted.json
@@ -1,105 +1,0 @@
-{
-  "test_bool": true,
-  "test_int_32": -32,
-  "test_uint_32": 32,
-  "test_double": 32.34,
-  "test_string": "lorem-ipsum",
-  "test_array": [
-    "one",
-    "two",
-    "three",
-    4,
-    5,
-    true
-  ],
-  "out": {
-    "dump": {
-      "out": true,
-      "conversations": true,
-      "on_begin": true,
-      "on_end": true
-    }
-  },
-  "on_begin": {
-    "function": "OnScenarioBegin"
-  },
-  "on_end": {
-    "function": "OnScenarioEnd"
-  },
-  "conversations": [
-    {
-      "host": "localhost:7480",
-      "auth": {
-        "access_key": "test",
-        "secret_key": "test"
-      },
-      "out": {
-        "dump": {
-          "requests": true
-        }
-      },
-      "on_begin": {
-        "function": "OnConversationBegin"
-      },
-      "on_end": {
-        "function": "OnConversationEnd"
-      },
-      "requests": [
-        {
-          "for": 1,
-          "on_begin": {
-            "function": "OnRequestBegin"
-          },
-          "on_end": {
-            "function": "OnRequestEnd"
-          },
-          "auth": "aws_v4",
-          "method": "GET",
-          "uri": "foo",
-          "query_string": "format=json",
-          "response": {
-            "out": {
-              "format": {
-                "body": "json"
-              }
-            }
-          }
-        }
-      ]
-    },
-    {
-      "host": "localhost:7480",
-      "auth": {
-        "access_key": "test",
-        "secret_key": "test"
-      },
-      "out": {
-        "dump": {
-          "requests": false
-        }
-      },
-      "requests": [
-        {
-          "for": 1,
-          "auth": "aws_v2",
-          "method": "GET",
-          "uri": "admin/info",
-          "query_string": "format=json",
-          "response": {
-            "on_begin": {
-              "function": "OnResponseBegin"
-            },
-            "on_end": {
-              "function": "OnResponseEnd"
-            },
-            "out": {
-              "dump": {
-                "body": false
-              }
-            }
-          }
-        }
-      ]
-    }
-  ]
-}
