+++ test/test.h
@@ -1,11 +1,11 @@
 #pragma once
 
 #include "gtest/gtest.h"
-#include "chatterbox.h"
+#include "scenario.h"
 
 class cbox_test : public ::testing::Test {
   protected:
-    std::unique_ptr<rest::chatterbox> cbox;
+    std::unique_ptr<cbox::scenario> cbox;
     virtual void SetUp();
     virtual void TearDown();
 };
